AI Basics
Firsts steps in the AI implementation, with the MinMax algorithm
aplicated to this Tic-Tac-Toe game. At this point we don't have any
functionality for the two players mode, only a few features like the
menu and all the source code structure. We need to improve the AI to
achive a good opponent in Tic-Tac-Toe.