AI New Steps
We've sleek a little bit all the AI functionality, but we don't have a
basic game thinking yet. In fact, we've many things to improve, but
we've resolved an important problem, in relationship with our IDE
(Code::Blocks 10.05), and the way we executed the Release Version, and
the Debug Version. At this point, the problem is solved, the priorities
at the MinMax algorithm are correct (same numbers in the main function
and in the heuristics function) and the next step is to improve this AI.