AI Basic Functionality
At last, we've the basic functionality of the AI, with a logic game
thinking. At this point, the game is basically finished. From now, we've
to improve the AI reactions (we've a 1-level-MinMax algorithm, 'cause we
need to improve the board-comprovation functions, like three-in-a-row),
improve all the graphics stuff and the final release firts-contact.